+++ .vscode/c_cpp_properties.json
@@ -3,15 +3,19 @@
         "myDefaultIncludePath": "${workspaceFolder}/include"
     },
     "configurations": [
-    {
-        "name": "Linux",
-        "includePath": ["${myDefaultIncludePath}", "tester/**"],
-        "defines": [],
-        "compilerPath": "/usr/bin/c++",
-        "cStandard": "c11",
-        "cppStandard": "c++98",
-        "intelliSenseMode": "linux-clang-x64"
-    }
+        {
+            "name": "Linux",
+            "includePath": [
+                "${myDefaultIncludePath}",
+                "tester/**"
+            ],
+            "defines": [],
+            "compilerPath": "/usr/bin/c++",
+            "cStandard": "c11",
+            "cppStandard": "c++98",
+            "intelliSenseMode": "linux-clang-x64",
+            "configurationProvider": "ms-vscode.makefile-tools"
+        }
     ],
     "version": 4
 }
