--- tester/src/test_suite.cpp
@@ -1,11 +1,0 @@
-#include "vector.hpp"
-#include "libunitcpp.hpp"
-
-AUTO_TEST_SUITE(test)
-
-AUTO_TEST_CASE(buidi)
-{
-	std::cout << "ok let's go" << std::endl;
-}
-
-AUTO_TEST_SUITE_END()
